vdaf: Make verify key, nonce, and shard coins associated types
We treat these as `&[u8]` in order to avoid adding too many generics on
`Vdaf`. We use associated constants to define their expected length so
that we can sample them as needed. This means we need to add code to
make sure the length is correct.

We can avoid this by leaning more on the type system: make these
associated types instead. Where we needed, we add trait bounds to
express that we can run `Rng::gen()` to sample a key, nonce, or coins as
needed.